+++ app/Http/Livewire/Home.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Http\Livewire;
+
+use App\Models\Employee;
+use Livewire\Component;
+
+class Home extends Component
+{
+    public function render()
+    {
+        return view('livewire.home', [
+            'employees' => Employee::all()
+        ]);
+    }
+}
