+++ public/css/app.css
@@ -776,15 +776,15 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.-mx-3{
   margin-left: -0.75rem;
   margin-right: -0.75rem;
+}.my-28{
+  margin-top: 7rem;
+  margin-bottom: 7rem;
 }.-my-3{
   margin-top: -0.75rem;
   margin-bottom: -0.75rem;
 }.my-1{
   margin-top: 0.25rem;
   margin-bottom: 0.25rem;
-}.my-28{
-  margin-top: 7rem;
-  margin-bottom: 7rem;
 }.ml-3{
   margin-left: 0.75rem;
 }.-ml-px{
@@ -857,6 +857,10 @@ Ensure the default browser behavior of the `hidden` attribute.
   margin-top: -1px;
 }.-mr-6{
   margin-right: -1.5rem;
+}.mb-8{
+  margin-bottom: 2rem;
+}.mb-12{
+  margin-bottom: 3rem;
 }.block{
   display: block;
 }.inline-block{
@@ -899,6 +903,8 @@ Ensure the default browser behavior of the `hidden` attribute.
   height: 9rem;
 }.h-96{
   height: 24rem;
+}.h-80{
+  height: 20rem;
 }.h-14{
   height: 3.5rem;
 }.h-\[4rem\]{
@@ -909,6 +915,12 @@ Ensure the default browser behavior of the `hidden` attribute.
   height: 1.75rem;
 }.h-screen{
   height: 100vh;
+}.h-24{
+  height: 6rem;
+}.h-28{
+  height: 7rem;
+}.h-32{
+  height: 8rem;
 }.max-h-60{
   max-height: 15rem;
 }.min-h-screen{
@@ -949,12 +961,16 @@ Ensure the default browser behavior of the `hidden` attribute.
   width: 6rem;
 }.w-52{
   width: 13rem;
+}.w-10\/12{
+  width: 83.333333%;
 }.w-7{
   width: 1.75rem;
 }.w-32{
   width: 8rem;
 }.w-80{
   width: 20rem;
+}.w-28{
+  width: 7rem;
 }.min-w-0{
   min-width: 0px;
 }.min-w-\[2rem\]{
@@ -1228,6 +1244,10 @@ Ensure the default browser behavior of the `hidden` attribute.
   border-width: 2px;
 }.border-0{
   border-width: 0px;
+}.border-8{
+  border-width: 8px;
+}.border-4{
+  border-width: 4px;
 }.border-b-2{
   border-bottom-width: 2px;
 }.border-l-4{
@@ -1323,6 +1343,9 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.bg-gray-300{
   --tw-bg-opacity: 1;
   background-color: rgb(209 213 219 / var(--tw-bg-opacity));
+}.bg-gray-400{
+  --tw-bg-opacity: 1;
+  background-color: rgb(156 163 175 / var(--tw-bg-opacity));
 }.bg-black\/50{
   background-color: rgb(0 0 0 / 0.5);
 }.bg-gray-900\/50{
@@ -1330,6 +1353,33 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.bg-gray-600{
   --tw-bg-opacity: 1;
   background-color: rgb(75 85 99 / var(--tw-bg-opacity));
+}.bg-slate-400{
+  --tw-bg-opacity: 1;
+  background-color: rgb(148 163 184 / var(--tw-bg-opacity));
+}.bg-slate-200{
+  --tw-bg-opacity: 1;
+  background-color: rgb(226 232 240 / var(--tw-bg-opacity));
+}.bg-slate-100{
+  --tw-bg-opacity: 1;
+  background-color: rgb(241 245 249 / var(--tw-bg-opacity));
+}.bg-pink-200{
+  --tw-bg-opacity: 1;
+  background-color: rgb(251 207 232 / var(--tw-bg-opacity));
+}.bg-pink-50{
+  --tw-bg-opacity: 1;
+  background-color: rgb(253 242 248 / var(--tw-bg-opacity));
+}.bg-slate-300{
+  --tw-bg-opacity: 1;
+  background-color: rgb(203 213 225 / var(--tw-bg-opacity));
+}.bg-slate-800{
+  --tw-bg-opacity: 1;
+  background-color: rgb(30 41 59 / var(--tw-bg-opacity));
+}.bg-gray-700{
+  --tw-bg-opacity: 1;
+  background-color: rgb(55 65 81 / var(--tw-bg-opacity));
+}.bg-pink-700{
+  --tw-bg-opacity: 1;
+  background-color: rgb(190 24 93 / var(--tw-bg-opacity));
 }.bg-opacity-25{
   --tw-bg-opacity: 0.25;
 }.bg-cover{
@@ -1355,6 +1405,10 @@ Ensure the default browser behavior of the `hidden` attribute.
   padding: 0.75rem;
 }.p-0{
   padding: 0px;
+}.p-32{
+  padding: 8rem;
+}.p-12{
+  padding: 3rem;
 }.px-4{
   padding-left: 1rem;
   padding-right: 1rem;
@@ -1406,6 +1460,24 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.py-0{
   padding-top: 0px;
   padding-bottom: 0px;
+}.px-24{
+  padding-left: 6rem;
+  padding-right: 6rem;
+}.py-32{
+  padding-top: 8rem;
+  padding-bottom: 8rem;
+}.py-16{
+  padding-top: 4rem;
+  padding-bottom: 4rem;
+}.px-12{
+  padding-left: 3rem;
+  padding-right: 3rem;
+}.py-24{
+  padding-top: 6rem;
+  padding-bottom: 6rem;
+}.px-8{
+  padding-left: 2rem;
+  padding-right: 2rem;
 }.pt-6{
   padding-top: 1.5rem;
 }.pt-5{
@@ -1434,6 +1506,8 @@ Ensure the default browser behavior of the `hidden` attribute.
   padding-right: 2.25rem;
 }.pl-2{
   padding-left: 0.5rem;
+}.pr-10{
+  padding-right: 2.5rem;
 }.pr-8{
   padding-right: 2rem;
 }.pt-8{
@@ -1472,6 +1546,9 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.text-5xl{
   font-size: 3rem;
   line-height: 1;
+}.text-4xl{
+  font-size: 2.25rem;
+  line-height: 2.5rem;
 }.font-medium{
   font-weight: 500;
 }.font-semibold{
@@ -1558,6 +1635,12 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.text-gray-200{
   --tw-text-opacity: 1;
   color: rgb(229 231 235 / var(--tw-text-opacity));
+}.text-rose-400{
+  --tw-text-opacity: 1;
+  color: rgb(251 113 133 / var(--tw-text-opacity));
+}.text-cyan-600{
+  --tw-text-opacity: 1;
+  color: rgb(8 145 178 / var(--tw-text-opacity));
 }.underline{
   -webkit-text-decoration-line: underline;
           text-decoration-line: underline;
@@ -1675,6 +1758,8 @@ Ensure the default browser behavior of the `hidden` attribute.
   --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width) var(--tw-ring-offset-color);
   --tw-ring-shadow: var(--tw-ring-inset) 0 0 0 calc(1px + var(--tw-ring-offset-width)) var(--tw-ring-color);
   box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow, 0 0 #0000);
+}.focus-within\:ring-inset:focus-within{
+  --tw-ring-inset: inset;
 }.hover\:scale-125:hover{
   --tw-scale-x: 1.25;
   --tw-scale-y: 1.25;
@@ -1685,6 +1770,9 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.hover\:border-spa-100:hover{
   --tw-border-opacity: 1;
   border-color: rgb(178 179 214 / var(--tw-border-opacity));
+}.hover\:border-black:hover{
+  --tw-border-opacity: 1;
+  border-color: rgb(0 0 0 / var(--tw-border-opacity));
 }.hover\:bg-indigo-600:hover{
   --tw-bg-opacity: 1;
   background-color: rgb(79 70 229 / var(--tw-bg-opacity));
@@ -1812,6 +1900,10 @@ Ensure the default browser behavior of the `hidden` attribute.
 }.focus\:ring-1:focus{
   --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width) var(--tw-ring-offset-color);
   --tw-ring-shadow: var(--tw-ring-inset) 0 0 0 calc(1px + var(--tw-ring-offset-width)) var(--tw-ring-color);
+  box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow, 0 0 #0000);
+}.focus\:ring-0:focus{
+  --tw-ring-offset-shadow: var(--tw-ring-inset) 0 0 0 var(--tw-ring-offset-width) var(--tw-ring-offset-color);
+  --tw-ring-shadow: var(--tw-ring-inset) 0 0 0 calc(0px + var(--tw-ring-offset-width)) var(--tw-ring-color);
   box-shadow: var(--tw-ring-offset-shadow), var(--tw-ring-shadow), var(--tw-shadow, 0 0 #0000);
 }.focus\:ring-inset:focus{
   --tw-ring-inset: inset;
@@ -1959,6 +2051,9 @@ Ensure the default browser behavior of the `hidden` attribute.
   }.dark\:bg-gray-700{
     --tw-bg-opacity: 1;
     background-color: rgb(55 65 81 / var(--tw-bg-opacity));
+  }.dark\:bg-gray-600{
+    --tw-bg-opacity: 1;
+    background-color: rgb(75 85 99 / var(--tw-bg-opacity));
   }.dark\:bg-gray-900{
     --tw-bg-opacity: 1;
     background-color: rgb(17 24 39 / var(--tw-bg-opacity));
@@ -2125,6 +2220,8 @@ Ensure the default browser behavior of the `hidden` attribute.
     grid-template-columns: repeat(11, minmax(0, 1fr));
   }.sm\:grid-cols-12{
     grid-template-columns: repeat(12, minmax(0, 1fr));
+  }.sm\:flex-row{
+    flex-direction: row;
   }.sm\:items-start{
     align-items: flex-start;
   }.sm\:items-center{
