+++ resources/views/layouts/footer.blade.php
@@ -2,9 +2,12 @@
     <div class="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
         <div class="grid grid-cols-7 gap-3">
             <!-- Mission Statement -->
-            <div class="col-span-7 sm:col-span-2 p-6 bg-spa-200 flex flex-wrap content-center">
-                <div class="text-white">
-                    Beyond Esthetics means 'seeing beyond with what the natural eye sees'. Sometimes the most effective way to help a client is simply listening to a need or giving someone a hug. Knowing that the best treatment is sometimes good old-fashioned TLC, Tonya's husband Steve came up with the tagline, "Come get some TLC from TLG". Beyond Esthetics started with just Tonya and has grown to over 20 employees and two locations.
+            <div class="col-span-7 sm:col-span-2 p-6 bg-spa-200 flex flex-wrap content-center text-white">
+                <div>
+                    Beyond Esthetics means 'seeing beyond with what the natural eye sees'. Sometimes the most effective way to help a client is simply listening to a need or giving someone a hug. Knowing that the best treatment is sometimes good old-fashioned TLC, Tonya's husband Steve came up with the tagline, "Come get some TLC from TLG".
+                </div>
+                <div class="mt-4">
+                    Beyond Esthetics started with just Tonya and has grown to over 20 employees and two locations. Come see us today!
                 </div>
             </div>
             <!-- Contact Information -->
@@ -60,18 +63,26 @@
                 @endforeach
             </div>
         </div>
-        <div class="flex justify-between mt-3 border-t border-spa-200 py-3">
-            <div class="text-sm text-gray-600">&copy; {{ now()->format('Y') }} Beyond Esthetics</div>
-            <div class="flex flex-row items-center space-x-1">
+        <div class="flex flex-col sm:flex-row sm:justify-between sm:items-center mt-3 border-t border-spa-200 py-3">
+            <div class="text-sm text-gray-600 text-center">
+                &copy; {{ now()->format('Y') }} Beyond Esthetics
+            </div>
+            <div class="flex flex-row items-center space-x-1 mx-auto sm:mx-0">
                 <div class="text-sm text-gray-600">Designed with</div>
                 <span>
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" class="fill-current text-spa-200 h-3 w-3"><!--! Font Awesome Pro 6.0.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc.--><path d="M0 190.9v-5.8c0-69.9 50.52-129.52 119.4-141 44.7-7.59 92 7.27 124.6 39.92L256 96l11.1-11.98C300.6 51.37 347 36.51 392.6 44.1c68.9 11.48 119.4 71.1 119.4 141v5.8c0 41.5-17.2 81.2-47.6 109.5L283.7 469.1c-7.5 7-17.4 10.9-27.7 10.9s-20.2-3.9-27.7-10.9L47.59 300.4C17.23 272.1 0 232.4 0 190.9z"/></svg>
-                </span> 
+                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" class="fill-current text-spa-200 h-3 w-3">
+                        <!--! Font Awesome Pro 6.0.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc.-->
+                        <path d="M0 190.9v-5.8c0-69.9 50.52-129.52 119.4-141 44.7-7.59 92 7.27 124.6 39.92L256 96l11.1-11.98C300.6 51.37 347 36.51 392.6 44.1c68.9 11.48 119.4 71.1 119.4 141v5.8c0 41.5-17.2 81.2-47.6 109.5L283.7 469.1c-7.5 7-17.4 10.9-27.7 10.9s-20.2-3.9-27.7-10.9L47.59 300.4C17.23 272.1 0 232.4 0 190.9z" />
+                    </svg>
+                </span>
                 <div>+</div>
                 <span>
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"  class="fill-current text-red-600 h-3 w-3"><!--! Font Awesome Pro 6.0.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc.--><path d="M504.4 115.83a5.72 5.72 0 0 0-.28-.68 8.52 8.52 0 0 0-.53-1.25 6 6 0 0 0-.54-.71 9.36 9.36 0 0 0-.72-.94c-.23-.22-.52-.4-.77-.6a8.84 8.84 0 0 0-.9-.68L404.4 55.55a8 8 0 0 0-8 0L300.12 111a8.07 8.07 0 0 0-.88.69 7.68 7.68 0 0 0-.78.6 8.23 8.23 0 0 0-.72.93c-.17.24-.39.45-.54.71a9.7 9.7 0 0 0-.52 1.25c-.08.23-.21.44-.28.68a8.08 8.08 0 0 0-.28 2.08v105.24l-80.22 46.19V63.44a7.8 7.8 0 0 0-.28-2.09c-.06-.24-.2-.45-.28-.68a8.35 8.35 0 0 0-.52-1.24c-.14-.26-.37-.47-.54-.72a9.36 9.36 0 0 0-.72-.94 9.46 9.46 0 0 0-.78-.6 9.8 9.8 0 0 0-.88-.68L115.61 1.07a8 8 0 0 0-8 0L11.34 56.49a6.52 6.52 0 0 0-.88.69 7.81 7.81 0 0 0-.79.6 8.15 8.15 0 0 0-.71.93c-.18.25-.4.46-.55.72a7.88 7.88 0 0 0-.51 1.24 6.46 6.46 0 0 0-.29.67 8.18 8.18 0 0 0-.28 2.1v329.7a8 8 0 0 0 4 6.95l192.5 110.84a8.83 8.83 0 0 0 1.33.54c.21.08.41.2.63.26a7.92 7.92 0 0 0 4.1 0c.2-.05.37-.16.55-.22a8.6 8.6 0 0 0 1.4-.58L404.4 400.09a8 8 0 0 0 4-6.95V287.88l92.24-53.11a8 8 0 0 0 4-7V117.92a8.63 8.63 0 0 0-.24-2.09ZM111.6 17.28l80.19 46.15-80.2 46.18-80.18-46.17Zm88.25 60V278.6l-46.53 26.79-33.69 19.4V123.5l46.53-26.79Zm0 412.78L23.37 388.5V77.32L57.06 96.7l46.52 26.8v215.18a6.94 6.94 0 0 0 .12.9 8 8 0 0 0 .16 1.18 5.92 5.92 0 0 0 .38.9 6.38 6.38 0 0 0 .42 1 8.54 8.54 0 0 0 .6.78 7.62 7.62 0 0 0 .66.84c.23.22.52.38.77.58a8.93 8.93 0 0 0 .86.66l92.19 52.18Zm8-106.17-80.06-45.32 84.09-48.41 92.26-53.11 80.13 46.13-58.8 33.56Zm184.52 4.57L215.88 490.11V397.8l130.72-74.6 45.77-26.15Zm0-119.13L358.68 250l-46.53-26.79v-91.42l33.69 19.4L392.37 178Zm8-105.28-80.2-46.17 80.2-46.16 80.18 46.15Zm8 105.28V178L455 151.19l33.68-19.4v91.39Z"/></svg>
-                </span> 
+                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512" class="fill-current text-red-600 h-3 w-3">
+                        <!--! Font Awesome Pro 6.0.0 by @fontawesome - https://fontawesome.com License - https://fontawesome.com/license (Commercial License) Copyright 2022 Fonticons, Inc.-->
+                        <path d="M504.4 115.83a5.72 5.72 0 0 0-.28-.68 8.52 8.52 0 0 0-.53-1.25 6 6 0 0 0-.54-.71 9.36 9.36 0 0 0-.72-.94c-.23-.22-.52-.4-.77-.6a8.84 8.84 0 0 0-.9-.68L404.4 55.55a8 8 0 0 0-8 0L300.12 111a8.07 8.07 0 0 0-.88.69 7.68 7.68 0 0 0-.78.6 8.23 8.23 0 0 0-.72.93c-.17.24-.39.45-.54.71a9.7 9.7 0 0 0-.52 1.25c-.08.23-.21.44-.28.68a8.08 8.08 0 0 0-.28 2.08v105.24l-80.22 46.19V63.44a7.8 7.8 0 0 0-.28-2.09c-.06-.24-.2-.45-.28-.68a8.35 8.35 0 0 0-.52-1.24c-.14-.26-.37-.47-.54-.72a9.36 9.36 0 0 0-.72-.94 9.46 9.46 0 0 0-.78-.6 9.8 9.8 0 0 0-.88-.68L115.61 1.07a8 8 0 0 0-8 0L11.34 56.49a6.52 6.52 0 0 0-.88.69 7.81 7.81 0 0 0-.79.6 8.15 8.15 0 0 0-.71.93c-.18.25-.4.46-.55.72a7.88 7.88 0 0 0-.51 1.24 6.46 6.46 0 0 0-.29.67 8.18 8.18 0 0 0-.28 2.1v329.7a8 8 0 0 0 4 6.95l192.5 110.84a8.83 8.83 0 0 0 1.33.54c.21.08.41.2.63.26a7.92 7.92 0 0 0 4.1 0c.2-.05.37-.16.55-.22a8.6 8.6 0 0 0 1.4-.58L404.4 400.09a8 8 0 0 0 4-6.95V287.88l92.24-53.11a8 8 0 0 0 4-7V117.92a8.63 8.63 0 0 0-.24-2.09ZM111.6 17.28l80.19 46.15-80.2 46.18-80.18-46.17Zm88.25 60V278.6l-46.53 26.79-33.69 19.4V123.5l46.53-26.79Zm0 412.78L23.37 388.5V77.32L57.06 96.7l46.52 26.8v215.18a6.94 6.94 0 0 0 .12.9 8 8 0 0 0 .16 1.18 5.92 5.92 0 0 0 .38.9 6.38 6.38 0 0 0 .42 1 8.54 8.54 0 0 0 .6.78 7.62 7.62 0 0 0 .66.84c.23.22.52.38.77.58a8.93 8.93 0 0 0 .86.66l92.19 52.18Zm8-106.17-80.06-45.32 84.09-48.41 92.26-53.11 80.13 46.13-58.8 33.56Zm184.52 4.57L215.88 490.11V397.8l130.72-74.6 45.77-26.15Zm0-119.13L358.68 250l-46.53-26.79v-91.42l33.69 19.4L392.37 178Zm8-105.28-80.2-46.17 80.2-46.16 80.18 46.15Zm8 105.28V178L455 151.19l33.68-19.4v91.39Z" />
+                    </svg>
+                </span>
                 <div class="text-sm text-gray-600">by NT4C</div>
+            </div>
         </div>
-    </div>
 </footer>
