+++ resources/views/mx.blade.php
@@ -2,8 +2,8 @@
     <x-section title="Under Construction">
         <div class="bg-spa-100 p-3 rounded shadow my-28">
             <div class="text-lg font-mont text-center text-white">
-                Just finishing up the details on our new website. Sorry for the inconvenience. 
-                In the meantime please visit our social media channels below or call/email us with any questions!
+                <div>Just finishing up the details on our new website. Sorry for the inconvenience.</div>
+                <div>In the meantime, please visit our social media channels below or call/email us with any questions!</div>
             </div>
         </div>
     </x-section>
