+++ routes/web.php
@@ -1,7 +1,8 @@
 <?php
 
-use App\Http\Livewire\Welcome;
+use App\Http\Livewire\Home;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\ServiceController;
 
 /*
 |--------------------------------------------------------------------------
@@ -16,8 +17,12 @@ use Illuminate\Support\Facades\Route;
 
 Route::view('/', 'mx')->middleware('web')->name('mx');
 
-Route::get('/home', Welcome::class)->middleware('auth')->name('home');
+Route::middleware(['auth'])->group(function () {
 
-Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', function () {
-    return view('dashboard');
-})->name('dashboard');
+    // HOMEPAGE
+    Route::get('/home', Home::class)->name('home');
+
+    // SERVICES
+    Route::get('/services', [ServiceController::class, 'index'])->name('services');
+    Route::get('/services/{service:id}', [ServiceController::class, 'show'])->name('services.show');
+});
